pyproject.toml
[project]
name = "brainstorming-agent-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.1.4",
    "fastapi>=0.119.0",
    "firecrawl>=4.5.0",
    "httpx>=0.28.1",
    "langchain>=1.0.2",
    "langchain-openai>=1.0.1",
    "langchain-tavily>=0.2.12",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langgraph-cli[inmem]>=0.4.4",
    "numpy>=2.3.3",
    "psycopg[binary,pool]>=3.2.10",
    "python-dotenv>=1.1.1",
    "supabase>=2.20.0",
    "tavily-python>=0.7.12",
    "uvicorn>=0.38.0",
]
